[boxlogos] remove enigma-info dependency, but check for file existence

+++ meta-openpli/recipes-openpli/enigma2/enigma2-boxlogos.bb
@@ -4,8 +4,6 @@ HOMEPAGE = "https://github.com/DimitarCC/e2-boxlogos"
 require conf/license/license-gplv2.inc
 PACKAGE_ARCH = "${MACHINE_ARCH}"
 
-DEPENDS = "enigma-info"
-
 ALLOW_EMPTY:${PN} = "1"
 
 inherit gitpkgv ${PYTHON_PN}native
@@ -23,10 +21,14 @@ FILES:${PN} = "${datadir}/enigma2/logos/ ${libdir}/enigma2/python/Components/Ren
 FILES:${PN}-src = "${libdir}/enigma2/python/Components/Renderer/*.py"
 
 do_install() {
+	# make sure the enigma info file exists
+	INFOFILE=${DEPLOY_DIR_IMAGE}/../../enigma-info/${MACHINE}.txt
+	if [ ! -f $INFOFILE ]; then
+		exit 1
+	fi
+
 	install -d ${D}${datadir}/enigma2/logos
 
-	# get OE-A compatible build info
-	INFOFILE=${DEPLOY_DIR_IMAGE}/../../enigma-info/${MACHINE}.txt
 	MACHINE_BRAND=$( cat $INFOFILE | grep "displaybrand=" | cut -d "=" -f 2- )
 	DISTRO_NAME=$( cat $INFOFILE | grep "distro=" | cut -d "=" -f 2- )
 
@@ -78,7 +80,7 @@ fi
 #
 INFOFILE=/usr/lib/enigma.info
 if [ ! -f $INFOFILE ]; then
-    exit 0
+    exit 1
 fi
 
 # get the machine brand name for this image
